refactor: separate MaterialData and Material
