+++ Ignis/src/Ignis/Renderer/Material.h
@@ -8,26 +8,6 @@
 
 namespace ignis
 {
-	enum class MaterialType
-	{
-		Albedo,
-		Normal,
-		Metal,
-		Roughness,
-		Emissive,
-		AO
-	};
-
-	struct MaterialData
-	{
-		AssetHandle AlbedoMap = AssetHandle::Invalid;
-		AssetHandle NormalMap = AssetHandle::Invalid;
-		AssetHandle MetalnessMap = AssetHandle::Invalid;
-		AssetHandle RoughnessMap = AssetHandle::Invalid;
-		AssetHandle EmissiveMap = AssetHandle::Invalid;
-		AssetHandle AOMap = AssetHandle::Invalid;
-	};
-
 	class Material
 	{
 	public:
