+++ Ignis/src/Ignis/Scene/Components.h
@@ -2,7 +2,7 @@
 
 #include "SceneCamera.h"
 #include "Ignis/Core/UUID.h"
-#include "Ignis/Renderer/Material.h"
+#include "Ignis/Renderer/MaterialData.h"
 
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
